+++ packages/swapper/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@shapeshiftoss/swapper-v11.3.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v11.3.0...@shapeshiftoss/swapper-v11.3.1) (2022-09-27)
+
+
+### Bug Fixes
+
+* **swapper:** typo in MidgardCoins ([#1040](https://github.com/shapeshift/lib/issues/1040)) ([2c29b50](https://github.com/shapeshift/lib/commit/2c29b500f2d5a2dc12b11a810a46c70060da9b64))
+
+# [@shapeshiftoss/swapper-v11.3.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v11.2.0...@shapeshiftoss/swapper-v11.3.0) (2022-09-26)
+
+
+### Features
+
+* sellAssetAccountNumber -> bip44Params ([#1037](https://github.com/shapeshift/lib/issues/1037)) ([37ac813](https://github.com/shapeshift/lib/commit/37ac813ca997db0db5a9427d0e0eaadaa158b354))
+
+# [@shapeshiftoss/swapper-v11.2.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v11.1.3...@shapeshiftoss/swapper-v11.2.0) (2022-09-26)
+
+
+### Features
+
+* swap to/from RUNE ([#1030](https://github.com/shapeshift/lib/issues/1030)) ([6b5c9c8](https://github.com/shapeshift/lib/commit/6b5c9c8e855ffb68d865cfae8f545e7a819a9667))
+
 # [@shapeshiftoss/swapper-v11.1.3](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v11.1.2...@shapeshiftoss/swapper-v11.1.3) (2022-09-16)
 
 # [@shapeshiftoss/swapper-v11.1.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v11.1.1...@shapeshiftoss/swapper-v11.1.2) (2022-09-16)